+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Added support for IntelliJ platform version 2023.2
+- Added support for [`glhd/alpine-wizard`](https://github.com/glhd/alpine-wizard)
 
 ## [0.5.0] - 2023-05-31